Don't run every device in it's own thread
The fact that some device types need to be polled for new events is a
implementation detail, and we shouldn't force threads on all types of
devices, including for example CEC devices which work with callbacks
instead.

Thus, for Wiimotes, check for new events per device in the main
WiimoteController loop instead.