Revert "wiimote: port from xwiimote to wiiuse"
This partially reverts commit fea62f3aacd1097831046ce118d218742287cd81.

After talking to the authors of WiiUse, the library is not designed in a
way that we want. Basically it's meant to connect all your controllers
at once at the start of the program rather than continue finding remotes
in the background. It's not thread-safe to do otherwise.
This doesn't fit our use-case, where we don't know when the user
connects their Wiimote so we want to keep checking while checking for
events on already connected devices. Besides according to the authors
off WiiUse the kernel module approach actually has some benefits, and
after these discussions it does seem like the more solid approach.

The commit we're reverting did have some improvements though, for
example not starting a thread anymore in main.cpp and having a separate
thread to watch for events on all devices and a separate one to detect
Wiimotes. We're keeping those improvements